requirements.txt
typing-extensions==4.7.1
pathlib==1.0.1
python-dotenv==1.0.0
openai==1.3.0
gigachat==0.1.8
requests==2.31.0
aiohttp==3.8.5
numpy==1.24.3
pandas==2.0.3 